+++ plotoptix/install/__init__.py
@@ -1,0 +1,14 @@
+"""
+Install scripts for PlotOptiX extras.
+
+Documentation: https://plotoptix.rnd.team
+"""
+
+__author__  = "Robert Sulej, R&D Team <dev@rnd.team>"
+__status__  = "beta"
+__version__ = "0.4.0"
+__date__    = "05 July 2019"
+
+import logging
+
+logging.basicConfig(level=logging.WARN, format='[%(levelname)s] (%(threadName)-10s) %(message)s')