+++ plotoptix/tests/test_060_utilities.py
@@ -152,7 +152,9 @@ class TestScene(TestCase):
         self.assertTrue(np.min(w) >= -1 and np.max(w) <= 1, msg="Noise value out of <-1, 1> range.")
         self.assertFalse(np.isnan(w).any() or np.isinf(w).any(), msg="Noise contains NaNs or Infs.")
 
-    def test040_simplex(self):
+    def test050_gpu_config(self):
+
+        print(get_gpu_architecture())
 
         set_gpu_architecture("Compute_50")
         set_gpu_architecture("Auto")